% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubscraper--package.R
\docType{package}
\name{pubscraper-package}
\alias{pubscraper}
\alias{pubscraper-package}
\title{pubscraper: Scrape PubMed, Compile, and Export to csv}
\description{
This package automates PubMed queries through businessPubMed. Allows for looping queries of search terms through a list of specified journals, specifically targeting author information. This package also helps to compile search results and export them to csv.
}
\keyword{internal}
