% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape2csv.R
\name{scrape2csv}
\alias{scrape2csv}
\title{Scrape PubMed, Compile information, and Export to csv}
\usage{
scrape2csv(
  narrow,
  broad = NULL,
  operator = NULL,
  journals = NULL,
  start,
  end = NULL,
  title = NULL,
  outpath = NULL,
  newfolder = TRUE,
  raw = TRUE,
  clist = TRUE,
  alist = FALSE,
  plist = FALSE,
  jlist = FALSE,
  report = TRUE
)
}
\arguments{
\item{narrow}{Character vector of length 1: key search terms in double quotes using AND as the operator between terms 
all enclosed within parentheses}

\item{broad}{Character vector of length 1: key search terms in quotes using OR as the operator between terms
all enclosed within parentheses}

\item{operator}{Character string: Defines the boolean operator (AND, OR, NOT) between narrow and broad 
(AND is default)}

\item{journals}{Character vector of length j: journals to query search terms through a loop}

\item{start}{Character string: start publication date for the query YYYY, or YYYY/MM, or YYYY/MM/DD format}

\item{end}{Character string: end publication date for the query YYYY, or YYYY/MM, or YYYY/MM/DD format 
(default is 3000/12/31)}

\item{title}{Character string: title of the search,`used as prefix to output}

\item{outpath}{Character string: path for the output}

\item{newfolder}{If `TRUE`, a new folder is created for the output (default is TRUE)}

\item{raw}{If `TRUE`, raw query results are exported to csv (default is TRUE)}

\item{clist}{If `TRUE`, a list of author email contacts cleaned of duplicates is exported
(default is TRUE)}

\item{alist}{If `TRUE`, a list of authors resulting from the query is exported (default is TRUE)}

\item{plist}{If `TRUE`, a list of publication titles resulting from the query is exported (default is TRUE)}

\item{jlist}{If `TRUE`, and `journals` is not specified, a list of journals resulting from the query
is exported (default is FALSE)}

\item{report}{If `TRUE`, a query report is exported to csv (default is TRUE)}
}
\description{
Queries PubMed using the extract_PubMed_data from the businessPubMed package.
If a journal list is defined, the query will be performed iteratively through a loop.
Results are compiled and exported to csv.
}
\examples{
my.keywords = '("COVID-19" AND "coronavirus")'
my.keyphrases = '("ARDS" OR "acute respiratory distress syndrome")'
my.journals = c("JAMA Cardiol", "Allergy Asthma Clin Immunol", "N Engl J Med")
my.title = "Covid"
my.start = "2019"
scrape2csv(narrow = my.keywords, broad = my.keyphrases, operator = "OR", journals = my.journals, 
  start = my.start, title = my.title)
# searches (("COVID-19" AND "coronavirus") OR ("ARDS" OR "acute respiratory distress syndrome")) 
#   from 2019 on in 3 selected journals
}
